requirements.txt
pydantic
pydantic-settings
appdaemon
python-statemachine
nest-asyncio